+++ lib/language_pack/base.rb
@@ -17,7 +17,7 @@ class LanguagePack::Base
 
   VENDOR_URL = ENV["BUILDPACK_VENDOR_URL"] || "https://heroku-buildpack-ruby.s3.dualstack.us-east-1.amazonaws.com"
   ROOT_DIR = File.expand_path("../../..", __FILE__)
-  MULTI_ARCH_STACKS = ["heroku-24"]
+  AMD_ONLY_STACKS = ["heroku-22"]
   KNOWN_ARCHITECTURES = ["amd64", "arm64"]
 
   attr_reader :app_path, :bundler, :cache, :environment_name, :stack