+++ lib/language_pack/helpers/download_presence.rb
@@ -8,27 +8,27 @@
 #
 #    download = LanguagePack::Helpers::DownloadPresence.new(
 #      'ruby-3.1.7.tgz',
-#      stacks: ['heroku-22', 'heroku-24']
+#      stacks: ['heroku-22', 'heroku-24', 'heroku-26']
 #    )
 #
 #    download.call
 #
 #    puts download.exists? #=> true
-#    puts download.valid_stack_list #=> ['heroku-22', 'heroku-24']
+#    puts download.valid_stack_list #=> ['heroku-22', 'heroku-24', 'heroku-26']
 class LanguagePack::Helpers::DownloadPresence
-  # heroku-22 and heroku-24 have identical ruby versions supported
-  STACKS = ["heroku-22", "heroku-24"]
+  # heroku-22, heroku-24, and heroku-26 have identical ruby versions supported
+  STACKS = ["heroku-22", "heroku-24", "heroku-26"]
 
-  def initialize(file_name:, arch:, multi_arch_stacks:, stacks: STACKS)
+  def initialize(file_name:, arch:, amd_only_stacks:, stacks: STACKS)
     @file_name = file_name
     @stacks = stacks
     @fetchers = []
     @threads = []
     @stacks.each do |stack|
-      @fetchers << if multi_arch_stacks.include?(stack)
-        LanguagePack::Fetcher.new(LanguagePack::Base::VENDOR_URL, stack: stack, arch: arch)
-      else
+      @fetchers << if amd_only_stacks.include?(stack)
         LanguagePack::Fetcher.new(LanguagePack::Base::VENDOR_URL, stack: stack)
+      else
+        LanguagePack::Fetcher.new(LanguagePack::Base::VENDOR_URL, stack: stack, arch: arch)
       end
     end
   end